Fix evpn_vxlan_health script for 8325H
Since 8325 was introduced in 10.15, the previous call to
/rest/v10.13/system/vrfs/default/routes/[ip]/[subnet_mask]/nexthops
was failing on 8325H because it defaulted to using REST v10.15,
which has a different response format since 10.13.  This commit
updates the call to use REST v10.13 (since this script's minimum
version is 10.13.1000) so that the format of the response is the
same across all platforms.